#10833
For reasons best known to Microsoft's IE7 team, it appears that "50% 50%" background image on a floated element puts the image in the centre of the viewport, rather than the sane thing of the content block.

Sticking it to "top left" fixes it (a 11px by 11px image in an 11px by 11px block, either of the 4 corners and dead centre are technically equivalent)... but question begs why it needed fixing in the first place. @_@ "Standards compliant browser"... right.


git-svn-id: file:///svn/phpbb/trunk@7557 89ea8834-ac86-4346-8a33-228a782c2dd0