[ticket/9173] No longer limit scope of numbers we store in the config table on PostgreSQL and Firebird when using set_config_count().
Since we're using a VARCHAR(255) column to store the numbers we have to CAST
the varchar string to a type we can do maths on. Using int4 or integer as the
type however limits the scope to 4-byte-integer = 32-bit. Using DECIMAL(255, 0)
allows the 'full' scope of decimals in varchar(255).

PHPBB3-9173