[ticket/9132] Oracle CLOB support is broken, preventing storage of long strings
This reverts Oracle support to the state it was in prior to phpBB 3.0.6.  That is, storage of long strings works again (e.g. posts > 4 KB), but the database backup/restore functionality is broken.  We feel that the ability to store long strings is more important than the DB restore, since Oracle 10g itself provides tools for backing up databases.

PHPBB3-9132