[feature/auto-loading] Added autoloader to common and ensured compatability
The autoloader is registered in install without caching, since caching is set
up during installation. This provides the same functionality, performance is
not a concern during installation anyway.

In common.php the autoloader is instantiated after the cache is initialised and
is started with the standard cache backend used for other cached values.

A few places in the code using class_exists had to be updated to explicitly
avoid autoloading, since they are used to check whether a file needs to be
included which does not (yet) follow the naming scheme the autoloader follows.

PHPBB3-9682