[ticket/9822] Correct some style related ACP explain statements.
Remove "Additionally, if you have permission you can elect to remove the set
from the filesystem." because that functionality is not implemented.

Also remove unnecessary "When the imageset is deleted it is gone for good."
statement, "It is recommended that you first export your set for possible
future use." should be clear enough.

PHPBB3-9822