[ticket/10014] Clearly indicate fatal errors in file acm.
If acm_file cannot open the cache file for writing, it prints a message
to that effect and calls die(). The message itself does not indicate that
it is a fatal error, and someone seeing the message might expect that
inability to write to cache is not fatal.

Make it clear that the error is fatal by printing the word "Fatal" before
the message.

PHPBB3-10014