[ticket/9997] Fixed an inconsistency in the Moderator Control Panel
The order of the Approve / Disapprove buttons was inconsistent in the
Moderator Control Panel - while on the main page and the moderation
queue itself the Approve button was to the right of the Disapprove
button, in the post details the Approve button was to the left of the
Disapprove button.

This very simple edit simply switches the position of these two buttons
in the post details page in the mcp.

PHPBB3-9995