[ticket/7834] Topic time didn't update when first post was deleted
When the first post of a topic was deleted, the topic time didn't
update - it should have changed to the time of the next post.

This commit simply applies lefty74's patch posted in the ticket. It gets
the post time of the next post from the database, and updates the thread
accordingly.

This patch is not my work at all and all credits go to lefty74, I just
transferred it onto GitHub

PHPBB3-7834