[feature/new-tz-handling] Started on database updater changes.
The changes are tricky since we need to drop user_dst column from
users table, but we need it during the data migration process to
properly calculate effective timezones for all users.

The change here converts user_timezone to vchar and computes
timezone identifiers from the offsets. It uses database-specific
functions for building SQL conditionals and concatenations which
need to be implemented, probably in a separate ticket. As a result
the current code is not functional.

PHPBB3-9558