[ticket/10247] Removing attempt_id column from the 3.0.8 to 3.0.9-RC1 updater.
To make sure that this column (which was too small and unecessary) does not
remain on 3.0.9-RCX installations and boards that were updated to a 3.0.9 RC,
the 3.0.10-RC1 release must correctly drop the column after db_tools has been
corrected and fully reviewed. The current version is not capable of dropping
primary keys correctly on all supported DBMSs.

PHPBB3-10247