[ticket/10300] Always set group attributes for teampage and legend
The code in acp_groups.php removes unchanged attributes from the
group_attributes array, to prevent issues with reapplying avatar/rank and more.
This code causes problems with the new teampage/legend feature, because when
the attribute is not set, the group is removed from the feature. Therefore I
added an array with the keys that need to be set, when calling the function.

PHPBB3-10300