[feature/extension-manager] Load (A/U/M)CP modules from extensions
To avoid large bc breaking changes, modules in the old includes directory
structure still follow the same naming conventions. Modules in extensions
have to be placed in an xcp/ folder and need a _module suffix. The
corresponding info file is in the same directory but with an _info suffix.

PHPBB3-10323