[ticket/10390] Improve the jQuery CDN url generation function
Per p's comments, return the remote url from a variable instead of
using multiple returns. Also put the logic for creating Google's version
on its own line, and count the version number's dots instead of length
so it will be less likely to break if jQuery goes to version 1.10.

PHPBB3-10390