[feature/save-post-on-report] Saves the post reported when a report is creat
A problem when creating a report is that the report may be made and then
the post changed. After the changes, if the report is seen the message
it shows may not be correct to be reported. With these changes, when
a report is created the system copies the post that was reported and
shows it in the report details in the MCP

PHPBB3-10600