[ticket/10633] Stop leaking filename of attachments when thumbnail is requested
While filenames are chosen at random and there is no correlation between the
original filename and the new filesystem filename, there is a correlation
between filesystem filename and filesytem filename of thumbnails.

Adjust error message to no longer include the physical filename and make it
consistent with the error message that is shown when there is no attachment at
all.

This information was mostly useless for regular users (i.e. non-admins) anyway.

PHPBB3-10633