[ticket/11029] Cache obtain_cfg_items should return empty array on failure
continue was used where it should not have been, causing a fatal error

This file is loaded on every page to check if style.cfg has changed. If it
has not, the user is not affected, so if it does not exist, the user should
not be affected either.

PHPBB3-11029