[ticket/8796] Mark read links updated to include time() in url
Submitting the current time() allows us to mark only the topics
and forums read up until a certain time (when the user loaded
the page). This means that any new posts or topics posted between
when the user opened the page and clicked the link are still
shown as unread.

PHPBB3-8796