[ticket/11103] Add is_disabled column to notifications table
EXTENSION AUTHORS TAKE NOTE! This is to prevent errors with notifications
from extensions!

Set is_disabled to 1 for all your notifications when your extension is
disabled so they are ignored and do not cause errors.

When your extension is enabled again, set is_disabled to 0 and your
notifications will be working again.

PHPBB3-11103