[feature/soft-delete] Add new fields for topic/post counts for softdelete
We can not use the replies vs replies_real anymore, as we need to be able to
determinate whether the posts are unapproved or softdeleted. So we need to add
a new field and there by change the second one for consistency.
We also add the posts_* fields for forums, which are a missing feature.

PHPBB3-9567