[feature/avatars] Use correct names of input fields
The name of the avatar input fields seem to have been copied from the local
avatar (gallery avatar). Since the input fields have different names in
the template files of the remote and gravatar avatars this will not
properly default to the entered values.
Additionally, the focused driver wasn't correctly filtered causing use to
always default to no avatar.

PHPBB3-10018