[ticket/10954] Change behavior of marking topics/forums
It will now display a popup message for 3 seconds which will confirm the
taken action. The amount of DOM traversals have been significantly reduced
and jQuery.each is now used instead of for loops. Additionally, it is now
possible to click on the mark topics/forums read links without triggering
an AJAX error.

PHPBB3-10954