[ticket/11314] Cleanup core.js coding
function errorHandler() was used in returnHandler() before it was defined.
The unused variable alert has been removed from errorHandler(). Overuse of
the var keyword for the variable sign has been removed and the if
statement has been simplified. The definition of the variable i has been
moved outside of the for loop definition. The var keyword shouldn't be
used inside the definition of the for loop.

PHPBB3-11314