[ticket/9657] Remove AJAX from delete button and approval/restore form
When a post is being deleted and the user goes to the next page, the first
post of that next page is now on the page before. So the user performing
the action will skip the post. Until we found a good solution to fix this,
we just remove the AJAX from these features.
For more information see http://tracker.phpbb.com/browse/PHPBB3-11340

PHPBB3-9657