[ticket/11442] Use correct button class for ajaxified confirm_box
In commit 001572f the HTML code for the ajaxified confirm_box was moved
from overall_footer.html to confirm_body.html. While copying, the CSS
class of the "Yes" button was changed from button1 to button2. Due to the
fact that the phpbb.confirm() method uses the class button1 to check if
"Yes" was clicked, this broke the ajaxified confirm box in the ACP. With
this small patch the confirm boxes in the ACP should work properly again.

PHPBB3-11442