[ticket/11538] Simplify colour value check and remove support for '#'
The input length for the hex color is now limited to 6 characters and
the support for colors starting with a '#' has been dropped. The allowed
input length of 7 in prosilver seems to have been a relict from old ages
of phpBB3. In order to have proper support for correct checking of the
colour value, the new code was also ported to the ACP groups manage page.
The tests have been modified to reflect the changes to the behavior of
the color check. Tests for the ACP will follow.

PHPBB3-11538