[feature/twig] Able to set chain of namespaces to search for loadTemplate()
This is done so that when event template files are included, if they
include files themselves, that namespace is checked first, then __main__
is checked to include the correct template file.

Also, when template files are included from a particular namespace, this is
done so that the files from that namespace are included first, then the
main namespace is checked.

We may want to change this behavior in the future to allow choosing which
locations have priority, but for now, this is what I am doing to make sure
the behavior is simple and always the same.

PHPBB3-11598