[feature/twig] INCLUDEPHP behavior now supports local relative paths
As a last resort, now we use the Twig Loader to find the correct file to
include to (most specific file first, then parent styles). Also allows
using @namespace convention. This is ONLY done if the specified path is not
an absolute path AND the file does not exist relative to the phpBB root path.

PHPBB3-11598