[ticket/11723] Correctly redirect user to agreement page and let him leave
This patch consists of two changes. The first one will make sure that $agree
is correctly reset to 0 and the user redirected back to the agreement page
after changing the display language. Secondly, by reseting 'change_lang',
the user will be able to agree to the terms on the agreement page again.
The changed language will still be kept, as this is correctly saved in the
'lang' field that is passed to the ucp_register page.
The variable $agree has also been changed to be boolean. It is not used as
an integer anywere in the ucp_register file.

PHPBB3-11723