[ticket/10917] Variable used only once so delete it
The variable $this->unequal_version was only used once
and only to display the version the package updates to.
To display the version it updates to makes no sense when
the update files just aren't meant to update from the
current version. (It's already shown in an error message)
So I deleted the variable from there.

Furthermore the use of version_compare makes the variable
useless in that context which is why I deleted the variable
from the whole file and replaced it in the relevant if statement
with the old comparison.

PHPBB3-10917