[ticket/11792] Add variable 'lang_set_ext' to event core.user_setup
To allow extensions to add global language strings just like mods can, add the
'lang_set_ext' variable to the core.user_setup event. It requires an ext_name
to be specified as well as a lang_set, and loads the specified lang_set in the
context of the extension.

PHPBB3-11792