[ticket/11805] Allow specifying the page in the route for pagination
Instead of "app.php?controller=news&start=10" authors might prefer having
the pagination in the route. So the sample link from above would be the
same as "app.php?controller=news/page/2". With this patch that is easily
possible. What you do is, you specify the $base_url and $start_name as follows:

	$base_url = $controller_helper->url('news/page/%d');
	$start_name = '/page/%d'; // Part of the url that will be removed
				  // for page 1

Now this will generate:
	"news"		for start = 0
		"news/page/1" will work aswell, but will not be created
	"news/page/2"	for start = 10

PHPBB3-11805