[ticket/11809] Ensure code.js is first script included after jQuery
The file core.js sets up most of the AJAX and jQuery related framework for
phpBB. Due to this, it needs to be included before any other javascript file
in order to ensure that subsequent files can use the phpBB variables and
functions. Currently, it is however loaded with INCLUDEJS in overall_footer
which causes it to be loaded after every other javascript file except for the
style specific ajax.js. This causes every javascript file that is included
before core.js and is using the phpBB AJAX functions or variables to break.

PHPBB3-11809