[ticket/9550] Add the core.viewtopic_post_rowset_data event to viewtopic.php
To allow extra fields added to the query retrieving post data to be used for
setting template variables, they need to be added to the post rowset. This
commit adds the core.viewtopic_post_rowset_data event that enables extension
developers to modify the post rowset when the database row is still available.

PHPBB3-9550