[ticket/11912] Add content_guesser
The content_guesser will try to use the function mime_content_type()
if it's available. If that is not the case, the content_guesser will try to
guess the mimetype using the file extension of the supplied file.
Since this guesser will be registered after the other guessers, it will
be only used if the other guessers are not available.

PHPBB3-11912