[ticket/11973] Remove logic from tokens language array
Previously if a translator did no translate a token, it would not be
displayed. On the other hand, the translator could add additional tokens
which have no functionality.  Functionality is already hardcoded into
bbcode.php so we can that in the ACP aswell, in order to prevent namend
issues.

PHPBB3-11973