[ticket/12355] Topic Tools not fully updated when subscribing/bookmarking
The Topic Tools menu appears at the top of a topic page and again at the
bottom of a topic page.
When you use the Topic Tools menu to Bookmark or Subscribe to a Topic, the
menu item you chose is updated to reflect that, by switching to say
Unsubscribe topic or Remove from Bookmarks.
However, this change is not mirrored in the other Topic Tools menu at the
other location on the page.
Both Topic Tools menus should be updated during the AJAX event, so that
they always show the same options to the user.
It is solved by creating a data-update-all attribute which contains all
the class/ID that will update when topic tool is clicked

PHPBB3-12355