[ticket/12386] Add DEBUG_EXTRA again and use it for container creation
We are currently creating a new compiled container on every page load when
having DEBUG enabled. However, one might only have that enabled to be
presented with errors or for getting the page load stats.
This change will add the DEBUG_EXTRA constant again. It will be used for
choosing whether the compiled container should be created on every page load -
when defined as true - or just once after the cache is cleared.

PHPBB3-12386