[ticket/12371] Do not delete post related notifications when a post is deleted
When a post is marked edited and therefor marked as softdeleted/unapproved,
we should not delete the notifications, just like we can't delete the emails.
Links are out there anyway and just keeping the notification is fail save.

If we keep the notification we can just continue to update it when more posts
are added. Also when we delete a post, we don't have to go though the table
and find all related notifications to be able to delete them or even update
them, when they have responders set.

PHPBB3-12371