[ticket/12270] Correct confirm approval message for topics
When clicking on approving a topic's first post in the MCP to
approve the topic itself, the confirm box dialog asks whether
to approve the post or not, instead of asking to approve the
topic.
To achieve this fix, we just need to count the topics to be
approved and check if it's greater than zero.

PHPBB3-12270