[ticket/12400] Add viewforum event to modify topics data
The event allows extensions to use and/or modify topics list and/or
topics data before dumping out the viewforum page.
For extensions which operate with viewforum page, f.e. if the topics list
and data is needed to retrieve, add, modify additional topics info
before the output loop.

PHPBB3-12400