[ticket/12407] Allow changing of post_data, etc. at end of posting.php
The current event core.posting_modify_template_vars is just ran using dispatch
but it doesn't pass any data from posting.php to the listener. Because of that,
it's not possible to know anything from posting.php and therefore limits the
use cases of this event. This will change it to allow similar actions as with
the event core.modify_posting_parameters.

PHPBB3-12407