[ticket/12155] Remove star hack from padding-right property of .button.
This was in place for IE7 due to the use of :after for inserting the button
icon. IE7 does not support :after, so the hack removed the extra padding that
made room for the icon. We no longer use :after for the icon, so this is no
longer an issue... not to mention that IE7 is no longer supported.

PHPBB3-12155