[ticket/12553] Fix links in the last post column that are not clickable.
The relative position of the <dt> is causing it to sit atop of the <dd> thus
preventing the user from clicking the links. The <dl> has a relative position
in LTR, which takes care of containing the read/unread icon link that has an
absolute position. We no longer support IE6, so removing the tweak that sets
the <dl> position to static will make everything behave the same way in RTL as
in LTR.

PHPBB3-12553