[ticket/14261] Move the update of session informations to page_footer()
Currently, the unique way to disable the update of session_page is to pass
"false" to the parameter of session_begin(). This method is directly
called in app.php, so pages served from the routing system can't disable
the update of session informations.

By moving the update to page_footer, we can allow controllers to tell to
the session manager that we don't want to update the session infos.

PHPBB3-14261