[ticket/16990] Fix for the style template code in the post editor
The code present in the prosilver style template causes the wrong
checkbox to be displayed in the post editor if the combination
`f_delete`:true and `f_softdelete`:false is active in the user
context. With this combination, this effectively prevents a forum
post from being deleted via the editor options.

Corrected the relevant section and switched to Twig syntax

PHPBB3-16990