[ticket/17369] Add forum id to approve action url for redirection
Added the forum_id to the approve action link which ensures that when
a topic is permanently deleted, the forum_id is present so that it can
be used in the redirection link to go to the previous forum. This addresses
a bug in which the forum_id is not provided and after deleting a topic, the
user will be taken to a 404 page as the forum id fallbacks to 0.

PHPBB-17369