[ticket/17373] Remove extra separator and change all separators to solid lines
The quick links menu in the header navbar had an additional horizontal
line at the bottom which did not need to be there. Additionally the
separators for links that always display in the quick links menu had
dotted underlines rather than the solid underlines that were under the
links that only show up for mobile users when there's less screen space.

PHPBB-17373