[ticket/17376] Add aria label to quote reference link
Adds an aria label attribute to the quoted post reference link so
that screen readers can inform the user of what the link is used
for. Before this change the link was too vague and would not have
been well described to users of its purpose if they couldn't see it.

PHPBB-17376