[ticket/17589] Check confirm id
When a login fails the qa captcha is reset and creates a new `confirm_id`.
However, the `confirm_id` is never checked for changes causing the captcha to
believe it's still solved and thus no need to generate a new captcha.

This commit checks the known `confirm_id` against the `confirm_id` given by the
user, marking the captcha solved accordingly.

PHPBB-17589